refactor: extract reusable components to address SonarCloud code duplication issue
